refactor azureiot code
