+++ libs/azureiot/pxt.json
@@ -10,9 +10,6 @@
     ],
     "public": true,
     "dependencies": {
-        "core": "file:../core",
-        "esp32spi": "file:../esp32spi",
-        "net": "file:../net"
+        "mqtt": "file:../mqtt"
     }
-
 }
