+++ libs/mqtt/README.md
@@ -1,7 +1,3 @@
 # MQTT
 
 MQTT communication layer. A port of https://github.com/rovale/micro-mqtt for MakeCode.
-
-## Changes
-
-* Wrapping code in ``mqtt`` namespace
