+++ pxtarget.json
@@ -34,7 +34,6 @@
         "libs/jacdac-services",
         "libs/jacdac-game",
         "libs/datalogger",
-        "libs/mqtt",
         "libs/lora",
         "libs/power",
         "libs/controller",
@@ -44,7 +43,8 @@
         "libs/proximity",
         "libs/feather",
         "libs/esp32spi",
-        "libs/azureiot",
-        "libs/net"
+        "libs/net",
+        "libs/mqtt",
+        "libs/azureiot"
     ]
 }
